:bug: bug: fix aspect ratio for BannerItem in mobile mode
