+++ app/hooks/useColorDarkenLighten.ts
@@ -4,14 +4,14 @@ import tinycolor from 'tinycolor2';
 
 /**
  * This hook returns lighten or darken color of a poster image
- * @param posterPath the path to poster image
+ * @param imageUrl the path to poster image
  * @returns an object containing isDark, loading, error colorDarkenLighten
  */
-export default function useColorDarkenLighten(posterPath?: string) {
+export default function useColorDarkenLighten(imageUrl?: string) {
   const { isDark } = useTheme();
 
   const { data, loading, error } = useColor(
-    `https://api.allorigins.win/raw?url=${encodeURIComponent(posterPath || '')}`,
+    `https://api.allorigins.win/raw?url=${encodeURIComponent(imageUrl || '')}`,
     'hex',
     {
       crossOrigin: 'anonymous',
